comanda-backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.5
